+++ src/BreadType.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Sonjaturo\DuckmodeFilament;
+
+enum BreadType
+{
+    case White;
+    case Brown;
+    case Multigrain;
+    case GlutenFree;
+    case Pita;
+    case Turkish;
+    case Raisin;
+    case Dwarf;
+    case None;
+}