+++ src/FeederWidget.php
@@ -3,6 +3,7 @@
 namespace Sonjaturo\DuckmodeFilament;
 
 use Filament\Widgets\Widget;
+use Filament\Widgets\WidgetConfiguration;
 
 class FeederWidget extends Widget
 {
@@ -13,6 +14,7 @@ class FeederWidget extends Widget
     public string $audioAsset = '';
 
     public int $starvationRate = 1000;
+    public int $bread = 5;
 
     protected static ?array $options = [
         'rotation' => 270, // start angle in degrees
@@ -52,5 +54,13 @@ class FeederWidget extends Widget
             'hungry' => __('duckmode::duck.hungry'),
             'starving' => __('duckmode::duck.starving'),
         ];
+    }
+
+    public static function make(array $properties = []): WidgetConfiguration
+    {
+        $bread = (isset($properties['bread']) && $properties['bread'] instanceof Bread) ? $properties['bread'] : new Bread(BreadType::White);
+        $properties['bread'] = $bread->getHealthValue();
+
+        return parent::make($properties);
     }
 }